extract shared autoscaler event utils and bundle grouping
- move build_autoscaler_scheduling_update_rows to shared module
  at ray._common.observability.autoscaler_event_utils to fix
  wrong dependency direction (dashboard -> autoscaler internals)
- extract group_resource_bundle_dicts shared utility
- update autoscaler_events_storage to use shared utilities
- re-export from event_logger for backward compatibility
