+++ python/ray/autoscaler/v2/event_logger.py
@@ -4,6 +4,9 @@ import logging
 from collections import defaultdict
 from typing import TYPE_CHECKING, Dict, List, Optional
 
+from ray._common.observability.autoscaler_event_utils import (
+    build_autoscaler_scheduling_update_rows,  # noqa: F401 - re-exported
+)
 from ray._private.event.event_logger import EventLoggerAdapter
 from ray.autoscaler.v2.utils import ResourceRequestUtil
 from ray.core.generated.autoscaler_pb2 import (
@@ -22,129 +25,6 @@ if TYPE_CHECKING:
     from ray.core.generated.autoscaler_pb2 import AutoscalingState
 
 logger = logging.getLogger(__name__)
-_TERMINATION_CAUSE_REASON_MAP = {
-    TerminationRequest.Cause.OUTDATED: "outdated",
-    TerminationRequest.Cause.MAX_NUM_NODES: "max number of worker nodes reached",
-    TerminationRequest.Cause.MAX_NUM_NODE_PER_TYPE: (
-        "max number of worker nodes per type reached"
-    ),
-    TerminationRequest.Cause.IDLE: "idle",
-}
-
-
-def build_autoscaler_scheduling_update_rows(
-    *,
-    cluster_resources: Dict[str, float],
-    launch_actions: Optional[List[Dict]] = None,
-    terminate_actions: Optional[List[Dict]] = None,
-    infeasible_resource_requests: Optional[List[Dict]] = None,
-    infeasible_gang_resource_requests: Optional[List[Dict]] = None,
-    infeasible_cluster_resource_constraints: Optional[List[Dict]] = None,
-) -> List[Dict[str, str]]:
-    rows = []
-
-    if launch_actions:
-        for launch_action in launch_actions:
-            rows.append(
-                {
-                    "severity": "INFO",
-                    "log_to_logger": True,
-                    "message": (
-                        "Adding "
-                        f"{launch_action['count']} node(s) of type "
-                        f"{launch_action['instance_type']}."
-                    ),
-                }
-            )
-
-    if terminate_actions:
-        for terminate_action in terminate_actions:
-            cause = terminate_action["cause"]
-            cause_reason = _TERMINATION_CAUSE_REASON_MAP.get(cause, "unknown")
-            rows.append(
-                {
-                    "severity": "INFO",
-                    "log_to_logger": True,
-                    "message": (
-                        "Removing "
-                        f"{terminate_action['count']} nodes of type "
-                        f"{terminate_action['instance_type']} ({cause_reason})."
-                    ),
-                }
-            )
-
-    if launch_actions or terminate_actions:
-        num_cpus = cluster_resources.get("CPU", 0)
-        resize_message = f"Resized to {int(num_cpus)} CPUs"
-        if "GPU" in cluster_resources:
-            resize_message += f", {int(cluster_resources['GPU'])} GPUs"
-        if "TPU" in cluster_resources:
-            resize_message += f", {int(cluster_resources['TPU'])} TPUs"
-        rows.append({"severity": "INFO", "message": f"{resize_message}."})
-
-    if infeasible_resource_requests:
-        log_str = "No available node types can fulfill resource requests "
-        for idx, request in enumerate(infeasible_resource_requests):
-            log_str += (
-                f"{request['resources']}*{request['count']}"
-                if "count" in request
-                else str(request["resources"])
-            )
-            if idx < len(infeasible_resource_requests) - 1:
-                log_str += ", "
-
-            if request.get("label_constraints"):
-                selector_strs = []
-                for constraint in request["label_constraints"]:
-                    values = ",".join(constraint["values"])
-                    selector_strs.append(
-                        f"{constraint['label_key']} "
-                        f"{constraint['operator']} [{values}]"
-                    )
-                if selector_strs:
-                    log_str += (
-                        " with label selectors: [" + "; ".join(selector_strs) + "]"
-                    )
-
-        log_str += ". Add suitable node types to this cluster to resolve this issue."
-        rows.append({"severity": "WARNING", "message": log_str})
-
-    if infeasible_gang_resource_requests:
-        for gang_request in infeasible_gang_resource_requests:
-            log_str = (
-                "No available node types can fulfill "
-                f"placement group requests (detail={gang_request['details']}): "
-            )
-            requests_by_count = gang_request["bundles"]
-            for idx, request in enumerate(requests_by_count):
-                log_str += (
-                    f"{request['resources']}*{request['count']}"
-                    if "count" in request
-                    else str(request["resources"])
-                )
-                if idx < len(requests_by_count) - 1:
-                    log_str += ", "
-
-            log_str += (
-                ". Add suitable node types to this cluster to resolve this issue."
-            )
-            rows.append({"severity": "WARNING", "message": log_str})
-
-    if infeasible_cluster_resource_constraints:
-        for infeasible_constraint in infeasible_cluster_resource_constraints:
-            log_str = "No available node types can fulfill cluster constraint: "
-            resource_requests = infeasible_constraint["resource_requests"]
-            for idx, request in enumerate(resource_requests):
-                log_str += f"{request['request']}*{request['count']}"
-                if idx < len(resource_requests) - 1:
-                    log_str += ", "
-
-            log_str += (
-                ". Add suitable node types to this cluster to resolve this issue."
-            )
-            rows.append({"severity": "WARNING", "message": log_str})
-
-    return rows
 
 
 class AutoscalerEventLogger:
